django_drf/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg[binary]==3.2.3
